# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TontineChain API",
        "description": "Documentation de l'API TontineChain pour le Hackathon MIABE 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        },
        {
            "url": "https://tonnine-benin-backend.onrender.com",
            "description": "Serveur de Production"
        }
    ],
    "paths": {
        "/api/v1/ai/chat": {
            "post": {
                "tags": [
                    "IA"
                ],
                "summary": "Converser avec YAO (Assistant Intelligent)",
                "operationId": "0db37d50f93bd2808dd6e4c4dbbbfa04",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Quel est mon solde ?"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "fr"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse de l'assistant AI",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "assistant": {
                                            "type": "string",
                                            "example": "YAO"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ton solde est de 50 000 FCFA"
                                        },
                                        "audio_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "demo_notice": {
                                            "properties": {
                                                "is_simulation": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "MULTILINGUAL AI message"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Utiliser le token Bearer retourné par /api/v1/auth/verify-otp",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}